Created SVGIcon #511 #discussion_r184814597
