+++ src/scripts/components/VideoFormInfoBox.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react'
 import { List as ImmutableList } from 'immutable'
 import { Link } from 'react-router-dom'
 import styled from 'styled-components'
-import { SVGIcon } from 'components/foundations/Button'
+import SVGIcon from 'components/foundations/SVGIcon'
 import VideoRecord from 'records/VideoRecords'
 import Card from './structures/Card'
 import TextField from './widgets/forms/TextField'
