+++ src/scripts/components/structures/header/MainHeader.js
@@ -4,7 +4,8 @@ import { Link } from 'react-router-dom'
 import Blockies from 'react-blockies'
 
 import SearchInputContainer from 'containers/widgets/SearchInputContainer'
-import Button, { SVGIcon } from 'components/foundations/Button'
+import Button from 'components/foundations/Button'
+import SVGIcon from 'components/foundations/SVGIcon'
 import MainHeaderLogo from 'components/widgets/MainHeaderLogo'
 import MainNavigation from 'components/structures/header/MainNavigation'
 import { add0x } from 'utils/AppUtils'
