+++ src/scripts/components/widgets/SearchInput.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react'
 import { withRouter } from 'react-router-dom'
 import styled from 'styled-components'
 
-import Button, { SVGIcon } from 'components/foundations/Button'
+import Button from 'components/foundations/Button'
+import SVGIcon from 'components/foundations/SVGIcon'
 import { SEARCH_PATH } from 'constants/UrlConstants'
 
 import type { RouterHistory } from 'react-router-dom'
