+++ src/scripts/components/widgets/WalletInfo.js
@@ -6,7 +6,7 @@ import styled from 'styled-components'
 import paratii from 'utils/ParatiiLib'
 import PTIBalanceContainer from 'containers/widgets/PTIBalanceContainer'
 import CloseButton from 'components/foundations/buttons/CloseButton'
-import { SVGIcon } from 'components/foundations/Button'
+import SVGIcon from 'components/foundations/SVGIcon'
 import Popover from 'components/foundations/Popover'
 import Colors from 'components/foundations/base/Colors'
 import { POPOVER_PADDING, OVERLAY_BUTTONS_HEIGHT } from 'constants/UIConstants'
