+++ src/scripts/components/widgets/modals/ModalShowSeed.js
@@ -5,7 +5,8 @@ import styled from 'styled-components'
 import Colors from 'components/foundations/base/Colors'
 import Title from 'components/foundations/Title'
 import Text from 'components/foundations/Text'
-import Button, { SVGIcon } from 'components/foundations/Button'
+import Button from 'components/foundations/Button'
+import SVGIcon from 'components/foundations/SVGIcon'
 import RadioCheck from 'components/widgets/forms/RadioCheck'
 import { ModalContentWrapper, ModalScrollContent } from './Modal'
 import { MNEMONIC_KEY_TEMP, PASSWORD_TEMP } from 'constants/ParatiiLibConstants'
